Include encoded URL in file name
If the `x-url` header is present in the request, include the %-encoded
contents of that in the name of the file that is created for editing.
This will allow the editor to be configured to use a different mode
depending on what is being edited.

For example, I have added a file with the following to `~/.vim/ftdetect`
to set the filetype to markdown for editing content from stackoverflow.

    " Not using setf command, since the filetype will already be set to text
    " Need to override that
    au BufNewFile,BufRead chrome_%%*stackoverflow*.txt	setlocal filetype=mkd